+++ weave-js/src/components/Panel2/PanelTraceTree/PanelTraceTreeTrace.tsx
@@ -1,26 +1,15 @@
 import * as globals from '@wandb/weave/common/css/globals.styles';
-import {SpanKindType, SpanType} from '@wandb/weave/core/model/media/traceTree';
-import React, {ReactNode, useCallback, useMemo} from 'react';
+import {SpanType} from '@wandb/weave/core/model/media/traceTree';
+import React, {useCallback, useMemo} from 'react';
 import {Loader} from 'semantic-ui-react';
 import styled from 'styled-components';
 
 import {useUpdatingState} from '../../../hookUtils';
 import {useNodeValue} from '../../../react';
-import {AgentSVG, ChainSVG, LLMSVG, ToolSVG} from '../Icons';
 import * as Panel2 from '../panel';
 import {PanelFullscreenContext} from '../PanelComp';
 import {TooltipTrigger} from '../Tooltip';
-import {
-  agentColor,
-  agentTextColor,
-  chainColor,
-  chainTextColor,
-  llmColor,
-  llmTextColor,
-  MinimalTooltip,
-  toolColor,
-  toolTextColor,
-} from './common';
+
 import * as S from './lct.style';
 import {useTipOverlay} from './tipOverlay';
 import {useTimelineZoomAndPan} from './zoomAndPan';
@@ -30,6 +19,8 @@ import {
   layoutTree,
 } from './layout';
 import {treesToFlat} from './util';
+import {SpanTreeDetail} from './PanelTraceSpan';
+import {getSpanIdentifier, getSpanDuration, getSpanKindStyle} from './common';
 
 const inputType = {
   type: 'wb_trace_tree' as const,
@@ -41,53 +32,6 @@ type PanelTraceTreeTraceProps = Panel2.PanelProps<
   typeof inputType,
   PanelTraceTreeTraceConfigType
 >;
-
-type SpanKindStyle = {
-  color: string;
-  textColor: string;
-  label: string;
-  icon: ReactNode;
-};
-
-function getSpanKindStyle(kind?: SpanKindType): SpanKindStyle {
-  switch (kind) {
-    case 'CHAIN':
-      return {
-        color: chainColor,
-        textColor: chainTextColor,
-        label: `Chain`,
-        icon: <ChainSVG />,
-      };
-    case 'AGENT':
-      return {
-        color: agentColor,
-        textColor: agentTextColor,
-        label: `Agent`,
-        icon: <AgentSVG />,
-      };
-    case 'TOOL':
-      return {
-        color: toolColor,
-        textColor: toolTextColor,
-        label: `Tool`,
-        icon: <ToolSVG />,
-      };
-    case 'LLM':
-      return {
-        color: llmColor,
-        textColor: llmTextColor,
-        label: `LLM`,
-        icon: <LLMSVG />,
-      };
-    default:
-      return {
-        color: '#f3f3f3',
-        textColor: '#494848',
-        label: `Span`,
-        icon: <></>,
-      };
-  }
-}
 
 const PanelTraceTreeTrace: React.FC<PanelTraceTreeTraceProps> = props => {
   const nodeValue = useNodeValue(props.input);
@@ -203,17 +147,6 @@ export const TraceTreeSpanViewer: React.FC<{
       )}
     </S.TraceWrapper>
   );
-};
-
-const getSpanIdentifier = (span: SpanType) => {
-  return span.name ?? span._name ?? span.span_kind ?? 'Unknown';
-};
-
-const getSpanDuration = (span: SpanType) => {
-  if (span.end_time_ms && span.start_time_ms) {
-    return span.end_time_ms - span.start_time_ms;
-  }
-  return null;
 };
 
 const TooltipTriggerWrapper = styled.div`
@@ -364,161 +297,6 @@ export const SpanElementsInner: React.FC<{
 };
 
 export const SpanElements = React.memo(SpanElementsInner);
-
-const SpanTreeDetail: React.FC<{
-  span: LayedOutSpanType;
-}> = props => {
-  const {span} = props;
-  const kindStyle = getSpanKindStyle(span.span_kind);
-  const identifier = getSpanIdentifier(span);
-  const duration = getSpanDuration(span);
-
-  return (
-    <S.TraceDetailWrapper>
-      <S.SpanDetailWrapper>
-        <S.SpanDetailHeader>
-          <span>
-            <span
-              style={{
-                color: kindStyle.textColor,
-              }}>
-              {kindStyle.icon}
-            </span>
-            {identifier}
-          </span>
-          {duration != null && <S.DurationLabel>{duration}ms</S.DurationLabel>}
-        </S.SpanDetailHeader>
-        <S.SpanDetailTable>
-          <tbody>
-            {span.status_message != null && (
-              <DetailKeyValueRow
-                style={{
-                  color: span.status_code === 'ERROR' ? '#EB1C45' : undefined,
-                }}
-                label="Status Message"
-                value={span.status_message}
-              />
-            )}
-            {span.results != null && (
-              <>
-                {span.results.map((result, i) => {
-                  return (
-                    <React.Fragment key={i}>
-                      <tr>
-                        <S.SpanDetailSectionHeaderTd colSpan={2}>
-                          Result Set {i + 1}
-                        </S.SpanDetailSectionHeaderTd>
-                      </tr>
-                      {result.inputs != null && (
-                        <React.Fragment>
-                          <tr>
-                            <S.SpanDetailIOSectionHeaderTd colSpan={2}>
-                              Inputs
-                            </S.SpanDetailIOSectionHeaderTd>
-                          </tr>
-                          {Object.entries(result.inputs).map(
-                            ([key, value], j) => {
-                              return (
-                                <DetailKeyValueRow
-                                  key={j}
-                                  label={key}
-                                  value={value}
-                                />
-                              );
-                            }
-                          )}
-                        </React.Fragment>
-                      )}
-                      {result.outputs != null && (
-                        <React.Fragment>
-                          <tr>
-                            <S.SpanDetailIOSectionHeaderTd colSpan={2}>
-                              Outputs
-                            </S.SpanDetailIOSectionHeaderTd>
-                          </tr>
-                          {Object.entries(result.outputs).map(
-                            ([key, value], j) => {
-                              return (
-                                <DetailKeyValueRow
-                                  key={j}
-                                  label={key}
-                                  value={value}
-                                />
-                              );
-                            }
-                          )}
-                        </React.Fragment>
-                      )}
-                    </React.Fragment>
-                  );
-                })}
-              </>
-            )}
-            <tr>
-              <S.SpanDetailSectionHeaderTd colSpan={2}>
-                Metadata
-              </S.SpanDetailSectionHeaderTd>
-            </tr>
-            {span.span_id != null && (
-              <DetailKeyValueRow label="ID" value={span.span_id} />
-            )}
-            {span.span_kind != null && (
-              <DetailKeyValueRow label="Kind" value={span.span_kind} />
-            )}
-            {span.status_code != null && (
-              <DetailKeyValueRow label="Status" value={span.status_code} />
-            )}
-            {span.start_time_ms != null && (
-              <DetailKeyValueRow
-                label="Start Time"
-                value={'' + new Date(span.start_time_ms)}
-              />
-            )}
-            {span.end_time_ms != null && (
-              <DetailKeyValueRow
-                label="End Time"
-                value={'' + new Date(span.end_time_ms)}
-              />
-            )}
-            {span.child_spans != null && span.child_spans.length > 0 && (
-              <DetailKeyValueRow
-                label="Child Spans"
-                value={span.child_spans.length}
-              />
-            )}
-            {span.attributes != null &&
-              Object.entries(span.attributes).map(([key, value], i) => {
-                return <DetailKeyValueRow key={i} label={key} value={value} />;
-              })}
-          </tbody>
-        </S.SpanDetailTable>
-      </S.SpanDetailWrapper>
-    </S.TraceDetailWrapper>
-  );
-};
-
-const safeValue = (value: any) => {
-  return typeof value === 'string' ? value : JSON.stringify(value, null, 2);
-};
-
-const DetailKeyValueRow: React.FC<{
-  label: string;
-  value: any;
-  style?: React.CSSProperties;
-}> = props => {
-  const {label, value} = props;
-  const textValue = safeValue(value);
-  return (
-    <tr style={props.style}>
-      <S.KVDetailKeyTD>{'' + label}</S.KVDetailKeyTD>
-      <S.KVDetailValueTD>
-        <MinimalTooltip text={textValue}>
-          <S.KVDetailValueText>{textValue}</S.KVDetailValueText>
-        </MinimalTooltip>
-      </S.KVDetailValueTD>
-    </tr>
-  );
-};
 
 export const Spec: Panel2.PanelSpec = {
   id: 'wb_trace_tree-traceViewer',
